refactor: Ser has associated Error type now
